Fixes the biogenerator displaying item costs that were being divided by the machine's efficiency twice (#74939)
It was really bugging me that I would see a price being displayed in my
biogenerator, only for it to then cost me far more than what was
displayed. Looking into it, I realized that it was simply an issue with
it dividing the cost by the efficiency in the TGUI, when that math was
already done DM-side.

Being charged the cost you see on the machine is good.

:cl: GoldenAlpharex
fix: Biogenerators now display the accurate price of their products
after having been upgraded, as they no longer visually apply the
efficiency discount twice.
/:cl: