Taking items from your inventory while flying over lava or a chasm will NOT set them on fire or drop them ANYMORE (#75899)
Read the title. The code was modified a little to allow the pickup
animation to play even if the item is being moved from movable to
movable, and cut copypasta.

This will fix #75081

:cl:
fix: Taking items from your inventory while flying over lava or a chasm
will NOT set them on fire or drop them into chasms ANYMORE
/:cl: