Named barrels (#75976)
![dreamseeker_qLlw7wAl94](https://github.com/tgstation/tgstation/assets/3625094/747f61a0-533f-4cc2-ac61-2901d2d3afb4)

Barrels are nice stationary reagent containers, but it's hard to use
many of them because they're non-transparent and to see the reagents
with science glasses, you need to open the lid.

Now you can name barrels to not mistake them.

You don't need to open the barrel to figure out what you're going to get
from the tap.

:cl:
qol: You can name barrels with a pen now, changing its icon
/:cl: