Small connections.js refactor (#76563)
I wanted to use this for another project and I saw the commented out
code and was a little disappointment (because I wanted to use it)

Review with scrutiny I'm not good at JS

- Refactored `connections.js` into `connections.tsx`
- Moved it into `common` interface folder
- Uncommented subway code, allows for elements to use it via passing in
a prop

<Details>

<Summary>Interfaces work </summary>

![image](https://github.com/tgstation/tgstation/assets/51863163/e043f9ed-59a1-4d13-8526-e3660a81fdaa)

![image](https://github.com/tgstation/tgstation/assets/51863163/25a2c12b-414e-433f-8e17-b1bebe19822f)

</details>

Just some more generally useful components

:cl: Melbert
refactor: Refactor of connections.js. Yeah that's it.
/:cl: