Multi-tile airlocks (#76732)
Makes the large glass airlock operate properly on both tiles instead of
being a single tile airlock with a larger icon.

The right side of the current large glass airlock isn't really an
airlock tile, so any procs that check for the airlock's turf don't work
on the right side of the airlock.

<details>
<summary>Video</summary>

After:

https://github.com/tgstation/tgstation/assets/83487515/5bd571b5-fc9c-4135-942b-b712ae9ec960

Before:

https://github.com/tgstation/tgstation/assets/83487515/e752c085-e663-468b-ac21-2ae5d230b4be

</details>

:cl: LT3
fix: Large airlocks are now a subtype and function on both tiles
/:cl:

---------

Co-authored-by: san7890 <the@san7890.com>
Co-authored-by: Jacquerel <hnevard@gmail.com>