Augments/Prosthetic limbs now spark when shot instead of bleeding (#78248)
Title.

The amount of sparks are based on the damage dealt by the projectile
with slight variation, with one spark representing 25 damage.

Also cleans up some code while I was there (not because melbert gave me
a side eye or anything for not doing it at first hahaahahah)

Bleeding metal is weird, but I feel that having some extra feedback for
getting shot was nice - it felt like something was missing when I was
testing it out... so I added the sparks.
:cl:
fix: Projectiles no longer cause a blood graphic or blood splatters if
they hit a limb that cant bleed
add: Prosthetics/Augments now spark when shot
/:cl:

---------

Co-authored-by: Jacquerel <hnevard@gmail.com>
Co-authored-by: Ghom <42542238+Ghommie@users.noreply.github.com>