Honorbound Sect Improvements: Creatures are not Innocent, and Neither is the Blatantly Evil. Declarations are Cheaper, too (#80048)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

Innocence is no longer considered when an honorbound attacks a nonhuman
(as in mob, not like, lizardpeople, lol).
Attacking a cultist with a halo, or a blatant syndicate operative will
now immediately make THEM guilty.
Converting a new follower now gives 300 favor, enough for 2 declarations
(costing 150 each)

Feedback and stats taken from
https://tgstation13.org/phpBB/viewtopic.php?p=713399#p713399
This is really the first half of fixing Honorbound Sect. It's the least
picked sect, because it gives a strange conditional pacifism and while
that's fun for roleplay, the pacifism is painfully difficult to work
with and there are just so many cases you have to ask out loud "HOW IS
THIS INNOCENT" when a xenomorph is running at you and you can't swing. I
still need to add rewards, but just for now let's make honorbound more
playable by putting aside lesser creatures as at least not innocent or
guilty (well they can be guilty, but it doesn't matter)

Once it's better to play with the honorbound rules, we can talk more
rewards like banners for fellow honorbound deacons and other stuff.

<!-- If your PR modifies aspects of the game that can be concretely
observed by players or admins you should add a changelog. If your change
does NOT meet this description, remove this section. Be sure to properly
mark your PRs to prevent unnecessary GBP loss. You can read up on GBP
and it's effects on PRs in the tgstation guides for contributors. Please
note that maintainers freely reserve the right to remove and add tags
should they deem it appropriate. You can attempt to finagle the system
all you want, but it's best to shoot for clear communication right off
the bat. -->

:cl:
balance: Honorbound no longer cares about innocence when it comes to
lesser creatures. They can still be considered unready in some cases.
balance: Attacking a cultist with a halo or a nuclear operative first
instantly makes THEM guilty, allowing further attacks.
balance: More favor for converting someone to the honorbound rules
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->