Moves `thinking_IC` variable to a trait (#80122)
This was another boolean that was used to just manage stuff codeside
that really was not accessed _too_ much and is ultimately not useful as
a variable on `/mob`. This just moves it to a trait because it's only
really used in a few spots for a similar intent+purpose.

Less stuff to deal with in the average view variables window whenever
looking at a mob, which is really nice and welcome.
Doesn't concern players.