Chasms prevent teleportation (#80102)
Chasms prevent teleportation

Fixes #80101

Being able to teleport people into chasms using beacons as an anchor is
bad. It has very little counterplay besides fishing out the beacon,
changing the teleport target (assuming it's not hidden), or killing
someone holding a handtele.

The alternate idea for this PR was to see if the chasm added some sort
of trait to the things falling in it and make it so the teleporter would
ignore things with that trait, but it seems like chasms don't do that.

:cl: Tattle
balance: you can no longer teleport into chasms
/:cl:

---------

Co-authored-by: tattle <article.disaster@gmail.com>