Legion Mob Abilities (#81082)
Converts legions hardcoded abilities to the cooldown action / mob
abilities system.

I also took the liberty of converting the hacky 360 second cooldowns in
a lot of the mob actions into simply disabling other abilities while the
ability is active, this will make it easier to have bosses without
everything on a shared cooldown and also not allow abilities to be used
simultaneously.

Paving the way for basic megafauna.

:cl:
refactor: Legions abilities have been changed into actions that can be
added to any mob.
/:cl:

---------

Co-authored-by: Changelogs <action@github.com>