[no gbp] fixes ai controller runtime (#81134)
fixes the flaky error that would happen in cleanbot controllers. i
discovered thru the profiler that a proc i was using to stop tracking
unreachable targets and caching them to a blacklist in selectbehavior
had a bit of a high real time cost, so if the mob got deleted midway
through for any reason would cause a runtime error. ive moved this proc
off selectbehavior and im instead letting the move loop cache this
target into the blacklist if it reaches the max pathing attempts
fixes a runtime that sometimes happens in ai controllers

:cl:
fix: fixes a runtime that sometimes happens in ai controllers
/:cl: