allows to reorder chat tabs & hides delete from main tab (#81455)
This pull request aims to hide the delete button from the main chat tab
as well as to allow reordering of the other chat tabs.
((Not to cause any issues with existing tabs, the variable has to be
true, so the hiding of the delete button only takes effect for new
players or when someone deleted all tabs once))

![grafik](https://github.com/tgstation/tgstation/assets/144968721/c1682cef-3e4f-4c4f-8394-bbf1345d4630)

![grafik](https://github.com/tgstation/tgstation/assets/144968721/ffe973a5-24eb-44ed-b8db-e3c1867935d1)
- I'm not quite sure, why the main tab has the delete button in the
first place, after all, it's not like the tab should be removed?
So, we can just hide the delete button on that tab and keep it always
there.
- Accidentally deleting a chat tab when one has multiple tabs set up
requires to change all tabs to the right to regain the previous order,
so why not simply allow to reorder all tabs except for the main tab.
(The main tab can neither be moved, nor can anything swapped with it)
:cl:
qol: hides the delete button on the main tab allows to reorder all other
chat tabs
/:cl: