Feast of Owls Heretic Ritual : Forsake your ascension (#83769)
This adds a simple ritual that let's a heretic forsake their ascension
for immediate power - 5 knowledge points.

Heretic suffers from lack of gimmicks and it's not really because of
lack of tools that could support them, but rather due to heretics going
straight for ascension (who can blame them), this ritual allows a
heretic to restrict themself from ascending and in exchange gain a
sizeable amount of research points. Not only that, but this will also
help with people learning heretic, as it allows a new player to jump
immediately into heretic without the struggle of early game (the
struggle of early-game justifies late-game power, by rejecting late-game
power heretic gets a jump-start with points.)

:cl:
add: Adds Feast of Owls ritual to heretic which allows one to forsake
their ascension in exchange for immediate power.
/:cl:

---------

Co-authored-by: carlarctg <53100513+carlarctg@users.noreply.github.com>