post-last minute revision
actually, lets just use a question mark, why not! who cares! you didnt see my PR in the time between this and the last commit, the MGS exclaimation joke would be funny but past that it wouldnt make much sense for a oblivious quirk to have a icon that says attention.