Hotfix to Convert Modsuits chameleon module
Currently the convert modsuit's chameleon modules do not function. After a bit of testing, it seems to be an issue with the fact the convert modsuit is the only modsuit in the game (from what Ive seen) equipable on the back and toolbelt. Saying the mod suit is prebuilt with a toolbelt storage module hardbuilt into it we can fix the chameleon module by limiting the modsuit to the toolbelt. At least until someone more insane than me tackles the chameleon module code.