we get nice things again as long as im not lazy
readded the power to fax pets/amogi (amogi part unintentional its just coded that way) BUT this time, you cant fax to centcom! trying to send a fax to centcom will fail. Also adds a admin subtype to faxes that come with hacking already AND notifies admins when it recieves letters