PDA ringtones now show a balloon alert (#87278)
This makes it so receiving a PDA message will give the holder a balloon
alert with their ringtone.

PDA messages are often ignored due to not being noticed, as many players
are more often focused on the main game screen (which runechat has
somewhat normalized) than the chat window. This makes things more
obvious that there's a PDA message.

:cl:
add: PDA ringtones now show a balloon alert to the PDA holder.
/:cl: