GPSes now show the general direction of crosslinked z-levels (#87437)
This makes the GPS UI give the general direction of a GPS on a different
linked z-level.

https://github.com/user-attachments/assets/98c6dfd8-5ced-4145-b14a-3813821ef30c

Makes navigating through space less of a chore, as previously, I believe
the only way was to manually write down or memorize what direction was
linked to what z-level.

:cl:
add: GPSes now show the general direction of cross-linked z-levels.
/:cl: