QOL: standardize patho equipment
Giving pathologists consistent equipment is great for understanding where they lay on the scale of things. As before they are very map-dependent, having two, one, or even none science goggles. Which dormants is kinda the expected way to find virus's to my understanding.

Syringes as given in tribute to that idea.

And if you take signer it completely deletes your gloves. Which is needed if you get a touch disease. So giving it to vendor is a healthy choice.