I FUCKIN FIXED IT
it used null as ai_turf, so it could never find the AI on that turf to find their VOX setting.  but the vox announcement grabbed the AIs turf, so that worked perfectly fine. now it get_turf's the AI, so itll give the AIs turf to get evaluated to find the AI on that turf, making this a entire waste of everyones time.