Adds a .NET 9 SDK setup action to linters (#88988)
OpenDream was recently bumped to .NET 9, which is not on our runner
image. I added an action which installs the required .NET version for
DMCompiler to function.

Adding this action is not only good as a quick hack fix, but also for
posterity.
I also considered the impact this has on our runner execution time, but
my hope is that it should not matter if we have .NET installed already.