Add a SS13.check_tick lua function (#89081)
## About The Pull Request

This adds a new function for Lua scripts: `SS13.check_tick()` - it acts
pretty much exactly like the `CHECK_TICK` macro in DM.

It accepts a single argument, a boolean, which if true - it will act
like `CHECK_TICK_HIGH_PRIORITY` instead.

## Why It's Good For The Game

this function is a repeated code pattern in lua scripts, so having it in
the SS13 lua library is useful.

## Changelog
:cl:
admin: Added a new function in Lua scripting, SS13.check_tick, to avoid
causing too much lag during loops and such.
/:cl: