Team antags no longer hide all of someone's other antags from the roundend report (#89555)
fixes https://github.com/tgstation/tgstation/issues/82128

if anyone was in an antag team, it'd remove _all of their antag datums_
from being listed in the roundend report - so if, say, someone was a
BB+traitor, they wouldn't appear as a traitor in the roundend report at
all.

this fixes that - instead, only antag datums associated with a team that
appeared in the report will be skipped.