Changes the paper mache robe to not use a verb and sleep for 3 seconds to handle a cooldown (#90438)
The robe now utilizes cooldown procs to limit casting, and calls procs
on the item itself to handle the summoning thing instead of a verb that
sleeps to handle a 'cooldown'.

This is like, ancient and we have stuff to handle this better now.
:cl:
code: The paper mache robe now works a bit better under the hood.
/:cl: