multitool found apc qol (#90896)
## About The Pull Request

Replaces the text description for the APC multitool search function with
directional arrows similar to Heretic.
(If someone has better ideas for a sprite, I'm ready to listen.)


https://github.com/user-attachments/assets/833b4f10-eab2-4833-bdf6-d02bd5820f48


## Why It's Good For The Game

The old search function was inconvenient for players, it looks much
better.
## Changelog
:cl:
qol: Replaces the text description for the APC multitool search function
with directional arrows similar to Heretic.
image: Sprite for multitool arrow

/:cl: