Firelocks now skip checking uninitialized turfs (#91221)
this makes it so `/obj/machinery/door/firedoor/proc/process_results()`
now returns early if the source turf is uninitialized, which usually
happens as a result of holodeck loading.

gets rid of an error, and processing uninitialized turfs feels like a
recipe for jank anyways

:cl:
fix: Firelocks now skip checking uninitialized turfs, which caused
errors with the holodeck.
/:cl: