Arachnid Sericulture
Arachnids are able to produce a silk material from their wrists to craft items and structures, at the cost of nutrition. This ability is tied to a silk gland organ.

Arachnids can now move through sticky webs without getting stuck

S blood bags are guaranteed to spawn in blood crates and can be made in the wall med-stations.

New species preview and description.